Remove deprecated ReflectionDBObject class
JAVA-3190
